+++ src/api/data-migration/component-data-migration.ts
@@ -325,13 +325,14 @@ export const doTheMigration = async (
 
     // Saving result with migrated version of stories into file
     await createAndSaveToFile({
-        ext: "cjs",
-        filename: `${from}---migrated`,
+        datestamp: true,
+        ext: "json",
+        filename: `${from}---${itemType}-to-migrate`,
         folder: "migrations",
         res: notNullMigratedItems,
     });
 
-    await modifyOrCreateAppliedMigrationsFile(migrationConfig);
+    await modifyOrCreateAppliedMigrationsFile(migrationConfig, itemType);
 
     if (itemType === "story") {
         await managementApi.stories.updateStories(
@@ -359,12 +360,14 @@ type SaveBackupStoriesToFile = ({
     folder,
     res,
 }: {
+    itemType: "story" | "preset";
     filename: string;
     folder: string;
     res: any;
 }) => Promise<void>;
 
-const saveBackupStoriesToFile: SaveBackupStoriesToFile = async ({
+const saveBackupToFile: SaveBackupStoriesToFile = async ({
+    itemType,
     res,
     folder,
     filename,
@@ -372,7 +375,7 @@ const saveBackupStoriesToFile: SaveBackupStoriesToFile = async ({
     await createAndSaveToFile({
         ext: "json",
         datestamp: true,
-        suffix: ".sb.stories",
+        suffix: itemType === "story" ? ".sb.stories" : ".sb.presets",
         filename,
         folder,
         res: res,
@@ -423,10 +426,11 @@ export const migrateProvidedComponentsDataInStories = async (
             });
         }
 
-        const backupFolder = path.join("backup", "stories");
+        const backupFolder = path.join("backup", itemType);
 
         // save stories to file as backup
-        await saveBackupStoriesToFile({
+        await saveBackupToFile({
+            itemType,
             filename: `before__${migrationConfig}__${from}`,
             folder: backupFolder,
             res: itemsToMigrate,