+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [5.0.7-beta.4](https://github.com/sb-mig/sb-mig/compare/v5.0.7-beta.3...v5.0.7-beta.4) (2023-06-15)
+
+
+### Bug Fixes
+
+* improve recognize migrations script ([e2c81d1](https://github.com/sb-mig/sb-mig/commit/e2c81d18c0eff545f77334227b15a829ee0f4426))
+
+## [5.0.7-beta.3](https://github.com/sb-mig/sb-mig/compare/v5.0.7-beta.2...v5.0.7-beta.3) (2023-06-15)
+
+
+### Bug Fixes
+
+* modify creation of file in between migration ([9d88b20](https://github.com/sb-mig/sb-mig/commit/9d88b207824e06b950c56b48e4c71377efb8c8dc))
+
+## [5.0.7-beta.2](https://github.com/sb-mig/sb-mig/compare/v5.0.7-beta.1...v5.0.7-beta.2) (2023-06-15)
+
+
+### Bug Fixes
+
+* make backups stories and presets before migrating both ([18ea8fd](https://github.com/sb-mig/sb-mig/commit/18ea8fd517bc32b9e8639563edf03d2c5a8c2232))
+
+## [5.0.7-beta.1](https://github.com/sb-mig/sb-mig/compare/v5.0.6...v5.0.7-beta.1) (2023-06-15)
+
+
+### Bug Fixes
+
+* update swc/core ([7eff0c5](https://github.com/sb-mig/sb-mig/commit/7eff0c543a4df1bb16aaca085b224e8aafe5475e))
+
 ## [5.0.6](https://github.com/sb-mig/sb-mig/compare/v5.0.5...v5.0.6) (2023-06-15)